Module

MAGEMinEnsemble.MAGEMinEnsemble — Module

Main module for MAGEMinEnsemble.jl.

FractionalCrystallisation Submodule

MAGEMinEnsemble.Crystallisation.bulk_crystallisation — Method

Out = bulk_crystallisation(T_array, P, bulk_init, database, oxides, max_steps, sys_in)

Perform a bulk crystallisation simulation using MAGEMin at constant intensive variables e.g., pressure, oxygen fugacity.

Inputs

T_array (Vector{Float64}): Temperatures in degrees Celsius.

P (Float64): Pressure in kbar.

bulk_init (Vector{Float64}): Initial bulk composition.

database: MAGEMin database for simulations. For example, "ig" is the igneous database of Holland et al., 2018.

oxides (Vector{String}): Oxides that correspond to values in the bulk composition. See MAGEMin documentation for accepted oxides.

sys_in (String): Unit for initial bulk composition, can be "wt" or "mol", for wt(defaults to "wt")

Outputs

Out (Vector{MAGEMinC.gminstruct{Float64, Int64}}): array of simulation outputs for each temperature step.

MAGEMinEnsemble.Crystallisation.fractional_crystallisation — Method

Out = fractional_crystallisation(T_array, P, bulk_init, database, oxides, max_steps, sys_in)

Perform a fractional crystallisation simulation using MAGEMin at constant intensive variables e.g., pressure, oxygen fugacity.

Inputs

T_array (Vector{Float64}): Temperatures in degrees Celsius.

P (Float64): Pressure in kbar.

bulk_init (Vector{Float64}): Initial bulk composition.

database: MAGEMin database for simulations.

oxides (Vector{String}): Oxides that correspond to values in the bulk composition. See MAGEMin documentation for accepted oxides.

sys_in (String): Unit for initial bulk composition, can be "wt" or "mol", for wt(defaults to "wt")

Outputs

Out (Vector{MAGEMinC.gminstruct{Float64, Int64}}): array of simulation outputs for each temperature step.

MonteCarloBulk Submodule

MAGEMinEnsemble.MonteCarloBulk.generate_bulk_mc — Method

generate_bulk_mc(bulk, abs_unc, n_samples)

Randomly sample normal distributions for each oxide defined by a bulk composition and its absolute uncertainty, n_samples times.

Inputs

bulk (AbstractDict{String, Float64}): Bulk composition with keys as oxide strings, e.g. SiO2.

abs_unc (AbstractDict{String, Float64}): Absolute uncertainties with keys that correspond to those in bulk.

n_samples (Int): number of times to sample normal distributions.

Outputs

bulk_mc (AbstractDict{String, Vector{Float64}}): Randomly sampled bulk composition.

InputValidation Submodule

MAGEMinEnsemble.InputValidation.check_buffer_if_offset — Method

check_buffer_if_offset(constant_inputs::OrderedDict, variable_inputs::OrderedDict)

If offset is provided, checks that buffer is also provided.

MAGEMinEnsemble.InputValidation.check_bulk_frac — Method

check_bulk_frac(td_database)

Check whether bulk_frac is either "bulk" or "frac".

MAGEMinEnsemble.InputValidation.check_constant_inputs_values — Method

check_constant_inputs_values(constant_inputs)

Check the contents of constant_inputs for typing, ensure that numeric values are Float64.

MAGEMinEnsemble.InputValidation.check_keys_oxygen — Method

oxides = check_keys_oxygen(od)

Returns keys in an ordered dictionary (od) that contain a capital "O".

MAGEMinEnsemble.InputValidation.check_matching_keys — Method

check_matching_keys(constant_inputs, variable_inputs)

Check no keys are defined in both constant and variable inputs.

MAGEMinEnsemble.InputValidation.check_required_inputs — Method

check_required_inputs(constant_inputs, variable_inputs)

Check that list of required inputs is defined in either constantinputs or variableinputs

MAGEMinEnsemble.InputValidation.check_td_database — Method

check_td_database(td_database)

Check whether the provided td_database flag is accepted by MAGEMin.

MAGEMinEnsemble.InputValidation.check_variable_inputs_vectors — Method

check_variable_inputs_vectors(variable_inputs)

Check that variable inputs are Vectors.

MAGEMinEnsemble.InputValidation.prepare_inputs — Method

prepare_inputs(constant_inputs, variable_inputs)

Prepare and validate both constantinputs and variableinputs prior to ensure they are suitable for MAGEMin.

Inputs:

constant_inputs (Dict): inputs that will remain unchanged between MAGEMin simulations.

variable_inputs (Dict): inputs that vary across MAGEMin simulations.

Outputs:

updated_constant_inputs (Dict): inputs that will remain unchanged between MAGEMin simulations.

updated_variable_inputs (Dict): inputs that vary across MAGEMin simulations.

MAGEMinEnsemble.InputValidation.replace_zero_oxides — Method

new_constant_inputs, new_variable_inputs = replace_zero_oxides!(constant_inputs, variable_inputs)

Replaces oxides, except H2O, defined as 0.0 with 0.001.

MAGEMinEnsemble.InputValidation.replace_zero_pressure — Method

new_constant_inputs, new_variable_inputs = replace_zero_pressure!(constant_inputs, variable_inputs)

Replaces pressures defined as 0.0 with 0.001

MAGEMinEnsemble.InputValidation.validate_buffer — Method

validate_buffer(constant_inputs, variable_inputs)

Ensure that provided buffer string(s) is permitted by MAGEMin.

MAGEMinEnsemble.InputValidation.validate_oxides — Method

validate_oxides(constant_inputs, variable_inputs)

Ensure that the defined bulk composition has the correct oxides included. Provides error messages that identify missing or extraneous oxides.

MAGEMinEnsemble.InputValidation.validate_positive_oxides — Method

validate_positive_oxides(constant_inputs, variable_inputs)

Ensures that defined oxide values are positive. Throws an error if they are negative.

MAGEMinEnsemble.InputValidation.validate_positive_pressure — Method

validate_positive_pressure(constant_inputs, variable_inputs)

Ensures that defined pressure value(s) are positive. Throws an error if they are negative.

GenerateEnsemble Submodule

MAGEMinEnsemble.GenerateEnsemble.create_T_array — Method

T_array = create_T_array(all_inputs)

Construct an array of temperatures from an initial temperature (Tstart), final temperature (Tstop) and temperature increment (Tstep). Ensure that the Tstep is the correct sign.

MAGEMinEnsemble.GenerateEnsemble.generate_output_filename — Method

output_file = generate_output_filename(variable_inputs)

Helper function to generate the output filename. Has different behaviour dependant on the number of oxides in variable_inputs.

MAGEMinEnsemble.GenerateEnsemble.get_bulk_oxides — Method

bulk_init, Xoxides = get_bulk_oxides(all_inputs)

Extract a bulk composition and matching oxide strings from all_inputs.

MAGEMinEnsemble.GenerateEnsemble.run_simulations — Function

Output = run_simulations(constant_inputs, variable_inputs, bulk_frac)

Generates and runs simulation ensembles from intensive variable grid. Extracts inputs from constant and variable_inputs, performs simulations, and saves the outputs to appropriately named .csv and metadata files.

Inputs

constant_inputs (OrderedDict): Intensive variables, excluding T, that remain constant across simulation ensemble.

variable_inputs (OrderedDict): Intensive variables, excluding T, that vary across the simulation ensemble.

bulk_frac (String): Flag to indicate whether bulk or fractional crystallisation simulations should be run. "bulk" indicates bulk crystallisation, "frac" indicates fractional crystallisation.

Keyword Arguments

sys_in (String): Indicate units for input bulk composition (defaults to "wt", wt.%). "mol" for mol.%.

output_dir (String): Path to save output directory to (defaults to current directory).

td_database (String): Flag indicating thermodynamic database to use (defaults to "ig": Green et al., 2025). See the MAGEMin github for options.

Outputs

results (Dict{String, Any}): simulation results, where keys are variable_input combinations.

MAGEMinEnsemble.GenerateEnsemble.setup_output_directory — Method

output_dir = setup_output_directory(output_dir)

Function to check whether specified output directory exists. If not, creates directory. If it does, checks for .csv files. If directory contains .csv files throws and error and tells the user to choose another new/empty directory.